Move SwipeMenuViewOptions into its own file
SwipeMenuView.swift had grown to 616 lines holding the options model, two
public protocols, the view itself, and three protocol conformances. The
options struct is a self-contained value type, so give it its own file.
No code changes beyond the move (trailing whitespace trimmed).
